+++ tomcat/src/main/java/org/apache/coyote/http11/Filter.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.http11.httpmessage.Request;
+import org.apache.coyote.http11.httpmessage.RequestURI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
