+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,6 +4,9 @@ import java.io.IOException;
 import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.httpmessage.Request;
+import org.apache.coyote.http11.httpmessage.Response;
+import org.apache.coyote.http11.handler.Handlers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
