+++ tomcat/src/main/java/org/apache/coyote/http11/RequestExtractor.java
@@ -6,6 +6,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
+import org.apache.coyote.http11.httpmessage.Request;
 
 public class RequestExtractor {
