+++ tomcat/src/main/java/org/apache/coyote/http11/handler/DefaultHandler.java
@@ -1,4 +1,8 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler;
+
+import org.apache.coyote.http11.ContentTypeParser;
+import org.apache.coyote.http11.httpmessage.Request;
+import org.apache.coyote.http11.httpmessage.Response;
 
 public class DefaultHandler extends Handler {
