+++ tomcat/src/main/java/org/apache/coyote/http11/handler/IndexHandler.java
@@ -1,12 +1,15 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler;
 
 import java.io.IOException;
+import org.apache.coyote.http11.ContentTypeParser;
+import org.apache.coyote.http11.httpmessage.Request;
+import org.apache.coyote.http11.httpmessage.Response;
 
-public class RegisterHandler extends Handler {
+public class IndexHandler extends Handler {
 
     @Override
     public Response handle(Request request) throws IOException {
-        String target = "register.html";
+        String target = "index.html";
 
         String resource = findResourceWithPath(target);
         String contentType = ContentTypeParser.parse(target);
