+++ tomcat/src/main/java/org/apache/coyote/http11/handler/RegisterHandler.java
@@ -1,13 +1,15 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler;
 
 import java.io.IOException;
+import org.apache.coyote.http11.ContentTypeParser;
+import org.apache.coyote.http11.httpmessage.Request;
+import org.apache.coyote.http11.httpmessage.Response;
 
-public class StaticResourceHandler extends Handler {
+public class RegisterHandler extends Handler {
 
     @Override
     public Response handle(Request request) throws IOException {
-        RequestURI requestURI = request.getRequestURI();
-        String target = requestURI.absolutePath();
+        String target = "register.html";
 
         String resource = findResourceWithPath(target);
         String contentType = ContentTypeParser.parse(target);
