+++ tomcat/src/main/java/org/apache/coyote/http11/httpmessage/Response.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httpmessage;
 
 public class Response {
